Extracts experimental features

+++ scripts/test-package.sh
@@ -1,27 +1,44 @@
 #!/usr/bin/env bash
 
-[[ -s "$HOME/.rvm/scripts/rvm" ]] && source "$HOME/.rvm/scripts/rvm"
+SKIP_PACK=0
+BUILD_DEBUG=0
+BUILD_RELEASE=1
+BUILD_IOS=1
+BUILD_ANDROID=1
+BUILD_NEW_ARCH=1
+BUILD_OLD_ARCH=1
 
 VERSION="$(jq -r '.version' packages/react-native-webgpu/package.json)"
 IOS_DESTINATION="OS=18.0,name=iPhone 16"
 
-set -x
 set -e
 
+exec 3>&1
+function print() {
+  echo "$1" >&3
+}
+exec &> "test-package.log"
+
+[[ -s "$HOME/.rvm/scripts/rvm" ]] && source "$HOME/.rvm/scripts/rvm"
+
 PRODUCTS_DIR="$(pwd)/products"
-rm -rf "${PRODUCTS_DIR}"
 mkdir -p "${PRODUCTS_DIR}"
 
 pushd packages/react-native-webgpu || exit 1
 
-rm -f "react-native-webgpu-${VERSION}.tgz"
-npm pack
+if [[ "$SKIP_PACK" != "1" ]]; then
+  print "Packing react-native-webgpu"
+  rm -f "react-native-webgpu-${VERSION}.tgz"
+  yarn tsc
+  npm pack
+fi
 
 popd || exit 1
 
 pushd examples/Example || exit 1
 
 function clean_native() {
+  print "Cleaning native code"
   pushd android || exit 1
   ./gradlew clean
   popd || exit 1
@@ -33,42 +50,88 @@ function build_ios() {
   xcodebuild -workspace Example.xcworkspace -scheme Example -configuration "$configuration" -destination "${IOS_DESTINATION}" -derivedDataPath build build | xcpretty
 }
 
-yarn remove react-native-webgpu
-rm -rf node_modules
-yarn add "../../packages/react-native-webgpu/react-native-webgpu-${VERSION}.tgz"
+if [[ "$SKIP_PACK" != "1" ]]; then
+  print "Installing react-native-webgpu from react-native-webgpu-${VERSION}.tgz package"
+  yarn remove react-native-webgpu
+  rm -rf node_modules
+  # Important, yarn caches local .tgz so delete it from the cache
+  rm -f "$(yarn config get cacheFolder)"/react-native-webgpu*
+  rm -f "$(yarn config get globalFolder)"/cache/react-native-webgpu*
+  yarn add "../../packages/react-native-webgpu/react-native-webgpu-${VERSION}.tgz"
+fi
 
-clean_native
+if [[ "$BUILD_OLD_ARCH" == "1" ]]; then
+  clean_native
 
-pushd ios || exit 1
-bundle exec pod install
-#build_ios Debug
-#mv build/Build/Products/Debug-iphonesimulator/Example.app "${PRODUCTS_DIR}/Example-Debug-oldarch.app"
-build_ios Release
-mv build/Build/Products/Release-iphonesimulator/Example.app "${PRODUCTS_DIR}/Example-Release-oldarch.app"
+  if [[ "$BUILD_IOS" == "1" ]]; then
+    pushd ios || exit 1
+    print "Installing pods for oldarch"
+    bundle exec pod install
+    if [[ "$BUILD_DEBUG" == "1" ]]; then
+      print "Building iOS oldarch Debug"
+      build_ios Debug
+      mv build/Build/Products/Debug-iphonesimulator/Example.app "${PRODUCTS_DIR}/Example-Debug-oldarch.app"
+    fi
+    if [[ "$BUILD_RELEASE" == "1" ]]; then
+      print "Building iOS oldarch Release"
+      build_ios Release
+      mv build/Build/Products/Release-iphonesimulator/Example.app "${PRODUCTS_DIR}/Example-Release-oldarch.app"
+    fi
+    popd || exit 1
+  fi
+
+  if [[ "$BUILD_ANDROID" == "1" ]]; then
+    pushd android || exit 1
+    if [[ "$BUILD_DEBUG" == "1" ]]; then
+      print "Building Android oldarch Debug"
+      ./gradlew assembleDebug -PnewArchEnabled=false
+      mv app/build/outputs/apk/debug/app-debug.apk "${PRODUCTS_DIR}/Example-Debug-oldarch.apk"
+    fi
+    if [[ "$BUILD_RELEASE" == "1" ]]; then
+      print "Building Android oldarch Release"
+      ./gradlew assembleRelease -PnewArchEnabled=false
+      mv app/build/outputs/apk/release/app-release.apk "${PRODUCTS_DIR}/Example-Release-oldarch.apk"
+    fi
+    popd || exit 1
+  fi
+fi
+
+if [[ "$BUILD_NEW_ARCH" == "1" ]]; then
+  clean_native
+
+  if [[ "$BUILD_IOS" == "1" ]]; then
+    pushd ios || exit 1
+    print "Installing pods for newarch"
+    RCT_NEW_ARCH_ENABLED=1 bundle exec pod install
+    if [[ "$BUILD_DEBUG" == "1" ]]; then
+      print "Building iOS newarch Debug"
+      build_ios Debug
+      mv build/Build/Products/Debug-iphonesimulator/Example.app "${PRODUCTS_DIR}/Example-Debug-newarch.app"
+    fi
+    if [[ "$BUILD_RELEASE" == "1" ]]; then
+      print "Building iOS newarch Release"
+      build_ios Release
+      mv build/Build/Products/Release-iphonesimulator/Example.app "${PRODUCTS_DIR}/Example-Release-newarch.app"
+    fi
+    popd || exit 1
+  fi
+
+  if [[ "$BUILD_ANDROID" == "1" ]]; then
+    pushd android || exit 1
+    if [[ "$BUILD_DEBUG" == "1" ]]; then
+      print "Building Android newarch Debug"
+      ./gradlew assembleDebug -PnewArchEnabled=true
+      mv app/build/outputs/apk/debug/app-debug.apk "${PRODUCTS_DIR}/Example-Debug-newarch.apk"
+    fi
+    if [[ "$BUILD_RELEASE" == "1" ]]; then
+      print "Building Android newarch Release"
+      ./gradlew assembleRelease -PnewArchEnabled=true
+      mv app/build/outputs/apk/release/app-release.apk "${PRODUCTS_DIR}/Example-Release-newarch.apk"
+    fi
+    popd || exit 1
+  fi
+fi
+
 popd || exit 1
 
-pushd android || exit 1
-#./gradlew assembleDebug -PnewArchEnabled=false
-#mv app/build/outputs/apk/debug/app-debug.apk "${PRODUCTS_DIR}/Example-Debug-oldarch.apk"
-./gradlew assembleRelease -PnewArchEnabled=false
-mv app/build/outputs/apk/release/app-release.apk "${PRODUCTS_DIR}/Example-Release-oldarch.apk"
-popd || exit 1
-
-clean_native
-
-pushd ios || exit 1
-RCT_NEW_ARCH_ENABLED=1 bundle exec pod install
-#build_ios Debug
-#mv build/Build/Products/Debug-iphonesimulator/Example.app "${PRODUCTS_DIR}/Example-Debug-newarch.app"
-build_ios Release
-mv build/Build/Products/Release-iphonesimulator/Example.app "${PRODUCTS_DIR}/Example-Release-newarch.app"
-popd || exit 1
-
-pushd android || exit 1
-#./gradlew assembleDebug -PnewArchEnabled=true
-#mv app/build/outputs/apk/debug/app-debug.apk "${PRODUCTS_DIR}/Example-Debug-newarch.apk"
-./gradlew assembleRelease -PnewArchEnabled=true
-mv app/build/outputs/apk/release/app-release.apk "${PRODUCTS_DIR}/Example-Release-newarch.apk"
-popd || exit 1
-
-popd || exit 1
+print "Products moved to ${PRODUCTS_DIR}"
